Occupation GET endpoint gemaakt. De blazor pagina aangepast zodat het met data wordt gevuld.

+++ RealDolmenInetum/ModelLibrary/Models/Occupation.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Type { get; set; }
-        public double Average_duration { get; set; }
+        public int Average_duration { get; set; }
         public string? Remark {  get; set; }
     }
 }
